get height at map is implemented

+++ src/dongo_entity/dongo_entity.rs
@@ -146,11 +146,12 @@ impl DongoEntity {
             && pos.y <= start.y.max(end.y)
     }
 
+    /// only implemented for DongoEntity::Terrain!
     pub fn get_height_at(&self, x: f32, y: f32) -> f32 {
         match self {
             DongoEntity::Terrain(_,_,_,dt_meta) => {
 
-                dt_meta.get_height_at(x as f64, y as f64) as f32
+                dt_meta.get_height_at(x,y) 
             },
             _ => panic!("get_height_at() not implemented for this entity type"),
         }
